backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncpg==0.30.0
click==8.3.0
colorama==0.4.6
Faker==37.11.0
fastapi==0.119.0
h11==0.16.0
httptools==0.7.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
PyYAML==6.0.3
redis==6.4.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
